+++ src/plone/app/mosaic/browser/static/js/mosaic.actions.js
@@ -384,6 +384,9 @@ define([
                 if(yes){
                     $.mosaic.selectLayout();
                 }
+            },
+            visible: function(){
+                return $.mosaic.options.available_layouts.length > 0;
             }
         });